+++ src/draw_line_bresenham.cpp
@@ -22,15 +22,15 @@ std::vector<cv::Point> get_line_points(cv::Point point_a, cv::Point point_b)
     {
         if (point_a.x > point_b.x)
         {
-            float temp_x = point_a.x;
+            int temp_x = point_a.x;
             point_a.x = point_b.x;
             point_b.x = temp_x;
-            float temp_y = point_a.y;
+            int temp_y = point_a.y;
             point_a.y = point_b.y;
             point_b.y = temp_y;
         }
-        float d_x = point_b.x - point_a.x;
-        float d_y = point_b.y - point_a.y;
+        int d_x = point_b.x - point_a.x;
+        int d_y = point_b.y - point_a.y;
         this_point.y = point_a.y;
         if (point_b.y > point_a.y) // đường thẳng đi lên y = ax + b;
         {
@@ -82,15 +82,15 @@ std::vector<cv::Point> get_line_points(cv::Point point_a, cv::Point point_b)
     {
         if (point_a.y > point_b.y)
         {
-            float temp_x = point_a.x;
+            int temp_x = point_a.x;
             point_a.x = point_b.x;
             point_b.x = temp_x;
-            float temp_y = point_a.y;
+            int temp_y = point_a.y;
             point_a.y = point_b.y;
             point_b.y = temp_y;
         }
-        float d_x = point_b.x - point_a.x;
-        float d_y = point_b.y - point_a.y;
+        int d_x = point_b.x - point_a.x;
+        int d_y = point_b.y - point_a.y;
         if (point_b.x > point_a.x)
         {
             this_point.x = point_a.x;
